admin_email is not correctly being pulled and used.
"admin_email" is unable to be retrieved because the Email class is not correctly namespaced.

In addition, the "admin_email" value is optionally an array containing the email and the name of the user (e.g. [ 'admin@example.com' => 'admin joe' ]). This will however fail the "filter_var" test for email. So, if the value is an array, just grab and use the email.